requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
groq>=0.5.0
instructor>=0.4.5
python-dotenv>=0.19.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
pydantic>=2.0.0